+++ vesper_cmcp/vsp_cmcp_datalist.h
@@ -11,6 +11,7 @@
 #define VSP_CMCP_DATALIST_H_INCLUDED
 
 #include <vesper_util/vsp_api.h>
+#include <stdint.h>
 
 #if defined __cplusplus
 extern "C" {
@@ -47,7 +48,7 @@ VSP_API int vsp_cmcp_datalist_free(vsp_cmcp_datalist *cmcp_datalist);
  * Returns non-zero and sets vsp_error_num() if failed.
  */
 VSP_API int vsp_cmcp_datalist_add_item(vsp_cmcp_datalist *cmcp_datalist,
-    int data_id, int data_length, void *data_pointer);
+    uint16_t data_id, uint16_t data_length, void *data_pointer);
 
 /**
  * Get pointer to data stored in the data list.
@@ -56,7 +57,7 @@ VSP_API int vsp_cmcp_datalist_add_item(vsp_cmcp_datalist *cmcp_datalist,
  * Returns NULL and sets vsp_error_num() if failed or length does not match.
  */
 VSP_API void *vsp_cmcp_datalist_get_data(vsp_cmcp_datalist *cmcp_datalist,
-    int data_id, int data_length);
+    uint16_t data_id, uint16_t data_length);
 
 #if defined __cplusplus
 }
